fix(solver): replay widening display provenance on memo hits
